Perf: Optimize bundle sizes and code-split heavy components on home/landing pages
- Remove moment dependency from compareDate.ts in favor of native Date timestamp comparisons to prevent bundling ~280 KB moment library into client bundles.

- Extract SwitchBoard into a standalone component and dynamically import it in _app.tsx so heavy dashboard layout and modal subtrees are never downloaded on static routes like / and /s.

- Dynamically import GroupsCard, MentorCard, and TasksCard in index.tsx so the home page navigation shell renders instantly without waiting for card subtrees.

- Reduces First Load JS for static routes from ~827 kB to ~486 kB (~40% reduction).
