+++ src/pages/_app.tsx
@@ -4,21 +4,20 @@ import "react-datepicker/dist/react-datepicker.css";
 
 import { ChakraProvider } from "@chakra-ui/react";
 import { AppProps } from "next/app";
-import { PropsWithChildren } from "react";
 import theme from "../theme";
 import Head from "next/head";
 import { trpcNext } from "../trpc";
 import { ToastContainer } from "react-toastify";
-import { SessionProvider, useSession } from "next-auth/react";
+import { SessionProvider } from "next-auth/react";
 import { useRouter } from "next/router";
-import invariant from "shared/invariant";
-import PageLoader from "components/PageLoader";
-import AppPageContainer from "components/AppPageContainer";
-import AppPage, { AppPageType } from "AppPage";
+import AppPage from "AppPage";
 import { isStaticPage, staticUrlPrefix } from "../static";
 import StaticPageContainer from "components/StaticPageContainer";
-import { loginCallbackUrl } from "shared/loginUrl";
-import { loginUrl } from "shared/loginUrl";
+import dynamic from "next/dynamic";
+
+// Dynamically import SwitchBoard so authenticated layout/modals
+// are not bundled into _app.js and downloaded on static routes.
+const SwitchBoard = dynamic(() => import("components/SwitchBoard"));
 import ErrorBoundary from "fundebug/ErrorBoundary";
 import "fundebug"; // Initialize Fundebug
 
@@ -90,51 +89,3 @@ function App({
 }
 
 export default trpcNext.withTRPC(App);
-
-function SwitchBoard({
-  children,
-  pageType,
-}: {
-  pageType?: AppPageType;
-} & PropsWithChildren) {
-  const { status } = useSession();
-  const router = useRouter();
-
-  // Invariant guaranteed by the caller
-  invariant(!isStaticPage(router.route), "non-static page");
-  const isAuthPage = router.route.startsWith("/auth/");
-
-  if (status == "loading") {
-    return <PageLoader />;
-  } else if (status == "unauthenticated") {
-    if (isAuthPage) {
-      return children;
-    } else if (router.asPath === "/") {
-      // Redirect to static page if the user attempts to access the home page.
-      void router.push(staticUrlPrefix);
-      return null;
-    } else {
-      // Redirect to login if they attempt to access specific sub-pages.
-      void router.push(loginUrl(router.asPath));
-      return null;
-    }
-  } else {
-    invariant(status == "authenticated", "session status");
-    if (isAuthPage) {
-      // Redirect to callbackUrl if specified. This is to avoid SSO IdP flashing
-      // the dashboard page as soon as the user logs in.
-      // Redirect to / if no callbackUrl is specified, e.g. when the user
-      // directly visits /auth/foo.
-      const url = loginCallbackUrl(router);
-      void router.replace(url);
-      return null;
-    } else if (router.route === "/oauth2/profile") {
-      // /oauth2/profile page doesn't need <AppPageContainer />
-      return children;
-    } else {
-      return (
-        <AppPageContainer pageType={pageType}>{children}</AppPageContainer>
-      );
-    }
-  }
-}
